try to cut down the doubling of RAM when writing to couchdb
some additional utility functions, and also an "async" bulksave
that isn't (yet) really async call to system2(curl), but instead
a call with a null reader and that expects a data frame as doc.
The data frame rows are converted to json as needed (10,000 at a time)
and simultaneously dropped from the passed df.  In theory, the memory hit
should be roughly just 10000 rows
 modified:   Rwork/couchUtils.R

call the async bulksave function
 modified:   Rwork/processPems.R