Seems to be working okay
modify bulk data storage so that it works this time.
also fixed a raft of bugs with the RJSONIO implementation
Also don't bother writing if a column is NA.
Also delete databases when doing a detector, so that old
runs get purged without doing it by hand via the CouchDB futon interface
 modified:   couchUtils.R

rejigger to use modified bulk data
 modified:   processPems.R

symlink to processPems.R to allow multiple runs.
Not sure if git stores symlinks properly
 new file:   processPems.b.R